requirements.txt
tb-mqtt-client
pyyaml
sensirion-i2c-sen5x
tabulate